backend/pyproject.toml
[tool.poetry]
name = "suppercx-backend"
version = "0.1.0"
description = ""
authors = ["Alvaro Acevedo <alvaro.acevedo.ing@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
fastapi = "^0.115.3"
crewai = { extras = ["tools"], version = "^0.76.2" }
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
uvicorn = "^0.24.0"
redis = "^5.2.0"
pydantic = "^2.9.2"
psycopg2-binary = "^2.9.10"
pytest-cov = "^5.0.0"
langchain-openai = "^0.2.3"
langchain-chroma = "^0.1.4"
voyageai = "^0.2.4"

[tool.poetry.group.dev.dependencies]
sqlalchemy-stubs = "^0.4"
black = "^24.10.0"
pylint = "^3.3.1"
pytest = "^8.3.3"

[tool.poetry.scripts]
dev = "uvicorn app.app:app --reload"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
